+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,27 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+import { fileURLToPath } from 'url';
+import { readFile } from 'fs/promises';
+
+const random = Math.random();
+
+export const unknownObject = (random > 0.5) ?
+  JSON.parse(await readFile('src/modules/files/a.json'))
+  :
+  JSON.parse(await readFile('src/modules/files/b.json'));
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+export const createMyServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
